fix: cleanup and change APY line msg
