--- src/__mocks__/__fixtures__/GetBoostSummary-2f9af2/28fa53.json
@@ -1,57 +1,0 @@
-{
-  "boostPools": {
-    "nodes": [
-      {
-        "asset": "Btc",
-        "feeTierPips": 5,
-        "boostShares": {
-          "aggregates": {
-            "sum": {
-              "fee": "880322",
-              "feeUsd": "594.163730978100000000000000000000",
-              "amount": "1762890095",
-              "amountUsd": "1189844.106415796300000000000000000000"
-            }
-          }
-        },
-        "apies": {
-          "nodes": [{ "projectedApy": "21.345600000000000000000000000000" }]
-        }
-      },
-      {
-        "asset": "Btc",
-        "feeTierPips": 10,
-        "boostShares": {
-          "aggregates": {
-            "sum": {
-              "fee": "0",
-              "feeUsd": "0",
-              "amount": "0",
-              "amountUsd": "0"
-            }
-          }
-        },
-        "apies": {
-          "nodes": [{ "projectedApy": "0.000000000000000000000000000000" }]
-        }
-      },
-      {
-        "asset": "Btc",
-        "feeTierPips": 30,
-        "boostShares": {
-          "aggregates": {
-            "sum": {
-              "fee": "0",
-              "feeUsd": "0",
-              "amount": "0",
-              "amountUsd": "0"
-            }
-          }
-        },
-        "apies": {
-          "nodes": [{ "projectedApy": "0.000000000000000000000000000000" }]
-        }
-      }
-    ]
-  }
-}
