--- src/__mocks__/__fixtures__/GetBoostSummary-2f9af2/4e0425.json
@@ -1,57 +1,0 @@
-{
-  "boostPools": {
-    "nodes": [
-      {
-        "asset": "Btc",
-        "feeTierPips": 5,
-        "boostShares": {
-          "aggregates": {
-            "sum": {
-              "fee": "21430595",
-              "feeUsd": "14458.526147320600000000000000000000",
-              "amount": "42892079944",
-              "amountUsd": "28937923.193108124600000000000000000000"
-            }
-          }
-        },
-        "apies": {
-          "nodes": [{ "projectedApy": "21.345600000000000000000000000000" }]
-        }
-      },
-      {
-        "asset": "Btc",
-        "feeTierPips": 10,
-        "boostShares": {
-          "aggregates": {
-            "sum": {
-              "fee": "0",
-              "feeUsd": "0",
-              "amount": "0",
-              "amountUsd": "0"
-            }
-          }
-        },
-        "apies": {
-          "nodes": [{ "projectedApy": "0.000000000000000000000000000000" }]
-        }
-      },
-      {
-        "asset": "Btc",
-        "feeTierPips": 30,
-        "boostShares": {
-          "aggregates": {
-            "sum": {
-              "fee": "0",
-              "feeUsd": "0",
-              "amount": "0",
-              "amountUsd": "0"
-            }
-          }
-        },
-        "apies": {
-          "nodes": [{ "projectedApy": "0.000000000000000000000000000000" }]
-        }
-      }
-    ]
-  }
-}
