+++ src/__mocks__/__fixtures__/GetBoostSummary-85840d/28fa53.json
@@ -7,15 +7,15 @@
         "boostShares": {
           "aggregates": {
             "sum": {
-              "fee": "1458435",
-              "feeUsd": "978.725312519400000000000000000000",
-              "amount": "2918721963",
-              "amountUsd": "1958695.647936658900000000000000000000"
+              "fee": "880322",
+              "feeUsd": "594.163730978100000000000000000000",
+              "amount": "1762890095",
+              "amountUsd": "1189844.106415796300000000000000000000"
             }
           }
         },
-        "apies": {
-          "nodes": [{ "projectedApy": "21.345600000000000000000000000000" }]
+        "apys": {
+          "nodes": [{ "projectedApy": "8.398500000000000000000000000000" }]
         }
       },
       {
@@ -31,7 +31,7 @@
             }
           }
         },
-        "apies": {
+        "apys": {
           "nodes": [{ "projectedApy": "0.000000000000000000000000000000" }]
         }
       },
@@ -48,7 +48,7 @@
             }
           }
         },
-        "apies": {
+        "apys": {
           "nodes": [{ "projectedApy": "0.000000000000000000000000000000" }]
         }
       }
